+++ h5_sdk/src/debug-adapter.ts
@@ -80,7 +80,9 @@ function sanitizeNetworkResponse(response: unknown): unknown {
     ...source,
     data: {
       echoAction: data.echoAction,
+      ok: data.ok,
       statusCode: data.statusCode,
+      statusText: data.statusText,
       duration: data.duration,
       headerNames: headers ? Object.keys(headers) : [],
       bodyLength: bodyText.length